combine image serializers into single file
