+++ api/v2/serializers/__init__.py
@@ -5,12 +5,11 @@ from .instance_serializer import InstanceSerializer
 from .instance_summary_serializer import InstanceSummarySerializer
 from .volume_serializer import VolumeSerializer
 from .volume_summary_serializer import VolumeSummarySerializer
-from .image_serializer import ImageSerializer
+from .image import ImageSerializer, ImageSummarySerializer, ImageBookmarkSerializer
 from .provider import ProviderSerializer, ProviderSummarySerializer, ProviderTypeSerializer, PlatformTypeSerializer
 from .identity import IdentitySerializer, IdentitySummarySerializer
 from .quota_serializer import QuotaSerializer
 from .allocation_serializer import AllocationSerializer
 from .provider_machine import ProviderMachineSerializer, ProviderMachineSummarySerializer
-from .image_bookmark_serializer import ImageBookmarkSerializer
 from .size_serializer import SizeSerializer
 from .size_summary_serializer import SizeSummarySerializer
