+++ api/v2/serializers/image.py
@@ -1,8 +1,8 @@
-from core.models import Application as Image
+from core.models import Application as Image, ApplicationBookmark as ImageBookmark
 from rest_framework import serializers
 from .user_serializer import UserSerializer
 from .tag_serializer import TagSerializer
-from .provider_machine_summary_serializer import ProviderMachineSummarySerializer
+from .provider_machine import ProviderMachineSummarySerializer
 
 
 class MachineProviderRelatedField(serializers.RelatedField):
@@ -17,21 +17,29 @@ class MachineProviderRelatedField(serializers.RelatedField):
 
 
 class ImageSerializer(serializers.HyperlinkedModelSerializer):
-    # todo: add created by field w/ user
-    # todo: add tags
     created_by = UserSerializer()
     tags = TagSerializer(many=True)
     provider_images = MachineProviderRelatedField(source='providermachine_set', many=True)
-    # machine_count = serializers.SerializerMethodField()
-    machine_count = serializers.IntegerField(
-        source='providermachine_set.count',
-        read_only=True
-    )
-
-    # def get_machine_count(self, obj):
-    #     return obj.providermachine_set.count()
 
     class Meta:
         model = Image
         view_name = 'api_v2:application-detail'
         fields = ('id', 'url', 'uuid', 'name', 'description', 'icon', 'created_by', 'tags', 'start_date', 'end_date', 'provider_images', 'machine_count')
+
+
+class ImageSummarySerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = Image
+        view_name = 'api_v2:application-detail'
+        fields = ('id', 'url', 'uuid', 'name', 'description', 'icon', 'start_date', 'end_date')
+
+
+class ImageBookmarkSerializer(serializers.HyperlinkedModelSerializer):
+    image = ImageSummarySerializer(source='application')
+    user = UserSerializer()
+
+    class Meta:
+        model = ImageBookmark
+        view_name = 'api_v2:applicationbookmark-detail'
+        fields = ('id', 'url', 'image', 'user')
